Move serializers and modify Watermark Function
